+++ src/main/java/classfit/example/classfit/exam/domain/enumType/ExamPeriod.java
@@ -1,0 +1,5 @@
+package classfit.example.classfit.exam.domain.enumType;
+
+public enum ExamPeriod {
+    WEEK, MONTH, DAILY, OTHER
+}
