+++ src/main/java/classfit/example/classfit/exam/domain/enumType/Standard.java
@@ -1,0 +1,5 @@
+package classfit.example.classfit.exam.domain.enumType;
+
+public enum Standard {
+    SCORE, QUESTION, PF, EVALUATION
+}
