+++ src/main/java/classfit/example/classfit/exam/domain/enumType/StandardStatus.java
@@ -1,0 +1,5 @@
+package classfit.example.classfit.exam.domain.enumType;
+
+public enum StandardStatus {
+    PASS,FAIL,EVALUATION,NONE
+}
