+++ src/main/java/classfit/example/classfit/exam/dto/exam/request/CreateExamRequest.java
@@ -1,0 +1,40 @@
+package classfit.example.classfit.exam.dto.exam.request;
+
+import classfit.example.classfit.course.domain.MainClass;
+import classfit.example.classfit.course.domain.SubClass;
+import classfit.example.classfit.exam.domain.Exam;
+import classfit.example.classfit.exam.domain.enumType.ExamPeriod;
+import classfit.example.classfit.exam.domain.enumType.Standard;
+import jakarta.validation.constraints.Min;
+import org.springframework.format.annotation.DateTimeFormat;
+
+import java.time.LocalDate;
+import java.util.List;
+
+public record CreateExamRequest(
+        Long subClassId,
+        Long mainClassId,
+        @DateTimeFormat(pattern = "yyyy-MM-dd")
+        LocalDate examDate,
+        Standard standard,
+        @Min(value = 0, message ="시험의 최고점수는 0이상 이어야 합니다.")
+        Integer highestScore,
+        ExamPeriod examPeriod,
+        String examName,
+        List<String> range
+) {
+
+    public Exam toEntity(final SubClass subClass, final MainClass mainClass) {
+        return Exam.builder()
+                .subClass(subClass)
+                .mainClass(mainClass)
+                .examName(examName)
+                .examDate(examDate)
+                .standard(standard)
+                .examPeriod(examPeriod)
+                .highestScore(highestScore)
+                .examRange(String.join(", ", this.range))
+                .build();
+    }
+
+}
