+++ src/main/java/classfit/example/classfit/exam/dto/exam/request/FindExamRequest.java
@@ -1,0 +1,9 @@
+package classfit.example.classfit.exam.dto.exam.request;
+
+public record FindExamRequest(
+        Long mainClassId,
+        Long subClassId,
+        String memberName,
+        String examName
+) {
+}
