+++ src/main/java/classfit/example/classfit/exam/dto/exam/response/FindExamStudentResponse.java
@@ -1,0 +1,11 @@
+package classfit.example.classfit.exam.dto.exam.response;
+
+import lombok.Builder;
+
+@Builder
+public record FindExamStudentResponse(
+        Long studentId,
+        String studentName
+) {
+
+}
