+++ src/main/java/classfit/example/classfit/scoreReport/service/ScoreReportService.java
@@ -22,8 +22,11 @@ import classfit.example.classfit.scoreReport.dto.request.SentStudentOpinionReque
 import classfit.example.classfit.scoreReport.dto.response.*;
 import classfit.example.classfit.student.domain.Student;
 import classfit.example.classfit.student.dto.StudentList;
-import classfit.example.classfit.studentExam.domain.*;
-import classfit.example.classfit.studentExam.dto.process.ExamHistory;
+import classfit.example.classfit.exam.domain.*;
+import classfit.example.classfit.exam.domain.enumType.Standard;
+import classfit.example.classfit.exam.dto.process.ExamHistory;
+import classfit.example.classfit.exam.repository.ExamRepository;
+import classfit.example.classfit.exam.repository.ExamScoreRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -43,7 +46,7 @@ public class ScoreReportService {
     private final SubClassRepository subClassRepository;
     private final ExamRepository examRepository;
     private final EnrollmentRepository enrollmentRepository;
-    private final StudentExamScoreRepository studentExamScoreRepository;
+    private final ExamScoreRepository examScoreRepository;
     private final AcademyRepository academyRepository;
 
     @Transactional(readOnly = true)
@@ -84,10 +87,10 @@ public class ScoreReportService {
             allStudents.add(new StudentList(report.getId(), student.getId(), student.getName()));
 
             for (Long examId : request.examIdList()) {
-                ExamScore examScore = studentExamScoreRepository.findByStudentAndExamId(student, examId)
+                ExamScore examScore = examScoreRepository.findByStudentAndExamId(student, examId)
                     .orElseThrow(() -> new ClassfitException(ErrorCode.SCORE_NOT_FOUND));
                 examScore.updateScoreReport(report);
-                studentExamScoreRepository.save(examScore);
+                examScoreRepository.save(examScore);
             }
         }
 
@@ -174,7 +177,7 @@ public class ScoreReportService {
     @Transactional
     public void deleteReport(@AuthMember Member member, Long studentReportId) {
         validateAcademy(member, member.getAcademy().getId());
-        studentExamScoreRepository.deleteByReportId(studentReportId);
+        examScoreRepository.deleteByReportId(studentReportId);
         scoreReportRepository.deleteById(studentReportId);
     }
 
@@ -248,7 +251,7 @@ public class ScoreReportService {
                 .mapToInt(AttendanceInfo::attendanceCount)
                 .sum();
 
-        List<ExamScore> examScores = studentExamScoreRepository.findByScoreReport(
+        List<ExamScore> examScores = examScoreRepository.findByScoreReport(
             scoreReport);
 
         List<ExamHistory> examHistoryList = examScores.stream()
@@ -258,8 +261,8 @@ public class ScoreReportService {
                 Exam exam = studentExamScore.getExam();
 
                     if (exam.getStandard() == Standard.PF) {
-                        long pCount = studentExamScoreRepository.countByExamAndScore(exam, -3);
-                        long fCount = studentExamScoreRepository.countByExamAndScore(exam, -4);
+                        long pCount = examScoreRepository.countByExamAndScore(exam, -3);
+                        long fCount = examScoreRepository.countByExamAndScore(exam, -4);
 
                         exam.updateAverage(pCount > fCount ? 100 : 0);
