--- src/main/java/classfit/example/classfit/studentExam/domain/ExamPeriod.java
@@ -1,5 +1,0 @@
-package classfit.example.classfit.studentExam.domain;
-
-public enum ExamPeriod {
-    WEEK, MONTH, DAILY, OTHER
-}
