--- src/main/java/classfit/example/classfit/studentExam/domain/Standard.java
@@ -1,5 +1,0 @@
-package classfit.example.classfit.studentExam.domain;
-
-public enum Standard {
-    SCORE, QUESTION, PF, EVALUATION
-}
