--- src/main/java/classfit/example/classfit/studentExam/domain/StandardStatus.java
@@ -1,5 +1,0 @@
-package classfit.example.classfit.studentExam.domain;
-
-public enum StandardStatus {
-    PASS,FAIL,EVALUATION,NONE
-}
