--- src/main/java/classfit/example/classfit/studentExam/dto/examRequest/FindExamRequest.java
@@ -1,9 +1,0 @@
-package classfit.example.classfit.studentExam.dto.examRequest;
-
-public record FindExamRequest(
-        Long mainClassId,
-        Long subClassId,
-        String memberName,
-        String examName
-) {
-}
