--- src/main/java/classfit/example/classfit/studentExam/dto/examResponse/FindExamStudentResponse.java
@@ -1,11 +1,0 @@
-package classfit.example.classfit.studentExam.dto.examResponse;
-
-import lombok.Builder;
-
-@Builder
-public record FindExamStudentResponse(
-        Long studentId,
-        String studentName
-) {
-
-}
