--- src/main/java/classfit/example/classfit/studentExam/dto/examScoreRequest/CreateExamScoreRequest.java
@@ -1,19 +1,0 @@
-package classfit.example.classfit.studentExam.dto.examScoreRequest;
-
-import classfit.example.classfit.studentExam.domain.Exam;
-import classfit.example.classfit.studentExam.domain.Standard;
-import classfit.example.classfit.studentExam.domain.StandardStatus;
-
-public record CreateExamScoreRequest(
-        Long studentId,
-        Long examId,
-        Integer score,
-        StandardStatus standardStatus,
-        String evaluationDetail,
-        boolean checkedStudent) {
-    public boolean isScoreRequired(Exam exam) {
-        return exam.getStandard() != Standard.PF && exam.getStandard() != Standard.EVALUATION;
-    }
-
-
-}
